[#22] 채팅 기능 구현 - 에러 코드 수정

+++ src/main/java/com/directors/domain/chat/ChatRepository.java
@@ -3,7 +3,7 @@ package com.directors.domain.chat;
 import java.util.List;
 
 public interface ChatRepository {
-    List<Chat> findChatListByRoomId(Long roomId, Long offset, Long size);
+    List<Chat> findChatListByRoomId(Long roomId, int offset, int size);
 
     void save(Chat chat);
 }
